third_party/blink/web_tests/http/tests/serviceworker/fetch-script-onerror-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: Uncaught ReferenceError: UndefinedReference is not defined
This is a testharness.js-based test.
[PASS] Check script error sanitization
Harness: the test ran to completion.
